requirements.txt
pandas==2.2.0
PyMuPDF==1.23.21
sentence-transformers==2.5.1
faiss-cpu==1.8.0
numpy==1.26.4
scikit-learn==1.4.1.post1
rank-bm25==0.2.2
python-dotenv==1.0.1
openai==1.13.3
streamlit==1.32.0
pytest==8.1.1
